apps/scanner/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scanner"
version = "0.1.0"
description = "PostgreSQL schema introspection and profiling service"
requires-python = ">=3.11"
dependencies = [
    "edna-common",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

